pyproject.toml
[project]
name = "diary"
version = "0.1.0"
description = "AI-powered daily task management TUI"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "pydantic>=2.0.0",
    "dateparser>=1.2.0",
    "blessed>=1.32.0",
]

[project.scripts]
diary = "diary.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/diary"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
